mcp-mysql-server/pyproject.toml
[project]
name = "mcp-mysql-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.3.0",
    "mysql-connector-python>=9.1.0",
]
